+++ js/i18n.js
@@ -1,36 +1,71 @@
 // global i18n object
 
 var i18n = new(function() {
+  function browser_locale() {
+    /* taken from https://github.com/maxogden/browser-locale by Max Ogden, BSD licensed */
+    var lang
+    
+    if (navigator.languages) {
+      // chrome does not currently set navigator.language correctly https://code.google.com/p/chromium/issues/detail?id=101138
+      // but it does set the first element of navigator.languages correctly
+      lang = navigator.languages[0]
+    } else if (navigator.userLanguage) {
+      // IE only
+      lang = navigator.userLanguage
+    } else {
+      // as of this writing the latest version of firefox + safari set this correctly
+      lang = navigator.language
+    }
+    
+    return lang
+  }
+
   var default_lng = "en";
-  var supported_lngs = [
+  var languages = {
     // translations found in locale/*.json
-    default_lng,
-    "ca",
-    "da",
-    "de",
-    "es",
-    "et",
-    "fr",
-    "hr",
-    "it",
-    "ja",
-    "nl",
-    "pt-BR",
-    "ru",
-    "sl",
-    "uk",
-    "vi",
-  ];
+    "en":    "English",
+    "ca":    "Catalan",
+    "da":    "Danish",
+    "de":    "German",
+    "es":    "Spanish",
+    "et":    "Estonian",
+    "fr":    "French",
+    "hr":    "Hungarian",
+    "it":    "Italian",
+    "ja":    "Japanese",
+    "nl":    "Dutch",
+    "no":    "Norwegian",
+    "pl":    "Polish",
+    "pt-BR": "Portuguese (Brazil)",
+    "ru":    "Russian",
+    "sl":    "Slovenian",
+    "uk":    "Ukrainian",
+    "vi":    "Vietnamese",
+    "zh-TW": "Chinese (Taiwan)"
+  };
+  var supported_lngs = _.keys(languages);
   this.getSupportedLanguages = function() {
-    return [].concat(supported_lngs);
+    return supported_lngs;
+  }
+  this.getSupportedLanguagesDescriptions = function() {
+    return languages;
   }
   this.getLanguage = function(lng) {
     lng = lng || settings.ui_language;
     if (lng == "auto") {
       // get user agent's language
       try {
-        lng = navigator.language.replace(/-.*/,"").toLowerCase();
+        lng = browser_locale().toLowerCase();
       } catch(e) {}
+      // hardcode some language fallbacks
+      if (lng === "nb") lng = "no"; // Norwegian Bokmål
+      // sanitize inconsistent use of lower and upper case spelling
+      var parts;
+      if (parts = lng.match(/(.*)-(.*)/))
+        lng = parts[1]+'-'+parts[2].toUpperCase();
+      // fall back to generic language file if no country-specific i18n is found
+      if ($.inArray(lng,supported_lngs) == -1)
+        lng = lng.replace(/-.*/,"");
     }
     return lng;
   }